requirements.txt
pocketflow>=0.0.1
PyYAML>=6.0
requests>=2.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0